# doc/api.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: Get the greeting message
          content:
            application/json:
              schema:
                type: string
      summary: Get the greeting message
      tags:
        - app
  /artist:
    post:
      operationId: ArtistController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: The artist has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad Request (Validation failure).
      summary: Create a new artist
      tags: &ref_0
        - artists
    get:
      operationId: ArtistController_findAll
      parameters: []
      responses:
        '200':
          description: Return all artists.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
      summary: Retrieve all artists
      tags: *ref_0
  /artist/{id}:
    get:
      operationId: ArtistController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The artist has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad Request (Validation failure).
        '404':
          description: Not Found.
      summary: Retrieve an artist
      tags: *ref_0
    put:
      operationId: ArtistController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: The artist has been successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad Request (Validation failure).
        '404':
          description: Not Found.
      summary: Update an artist
      tags: *ref_0
    delete:
      description: The artist has been successfully deleted.
      operationId: ArtistController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '400':
          description: Bad Request (Validation failure).
        '404':
          description: Not Found.
      summary: Delete an artist
      tags: *ref_0
  /album:
    post:
      operationId: AlbumController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: The album has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad Request (Validation failure).
      summary: Create a new album
      tags: &ref_1
        - albums
    get:
      operationId: AlbumController_findAll
      parameters: []
      responses:
        '200':
          description: Return all albums.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
      summary: Retrieve all albums
      tags: *ref_1
  /album/{id}:
    get:
      operationId: AlbumController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The album has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad Request (Validation failure).
        '404':
          description: Not Found.
      summary: Retrieve an album
      tags: *ref_1
    put:
      operationId: AlbumController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: The album has been successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad Request (Validation failure).
        '404':
          description: Not Found.
      summary: Update an album
      tags: *ref_1
    delete:
      description: The album has been successfully deleted.
      operationId: AlbumController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '400':
          description: Bad Request (Validation failure).
        '404':
          description: Not Found.
      summary: Delete an album
      tags: *ref_1
  /track:
    post:
      operationId: TrackController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: The track has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad Request (Validation failure).
      summary: Create a new track
      tags: &ref_2
        - tracks
    get:
      operationId: TrackController_findAll
      parameters: []
      responses:
        '200':
          description: Return all tracks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
      summary: Retrieve all tracks
      tags: *ref_2
  /track/{id}:
    get:
      operationId: TrackController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The track has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad Request (Validation failure).
        '404':
          description: Not Found.
      summary: Retrieve a track
      tags: *ref_2
    put:
      operationId: TrackController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: The track has been successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad Request (Validation failure).
        '404':
          description: Not Found.
      summary: Update a track
      tags: *ref_2
    delete:
      description: The track has been successfully deleted.
      operationId: TrackController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '400':
          description: Bad Request (Validation failure).
        '404':
          description: Not Found.
      summary: Delete a track
      tags: *ref_2
  /favs:
    get:
      operationId: FavoritesController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all favorites.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritesResponseDto'
      summary: Get all favorite artists, albums, and tracks
      tags: &ref_3
        - favorites
  /favs/track/{id}:
    post:
      operationId: FavoritesController_addTrack
      parameters:
        - name: id
          required: true
          in: path
          description: UUID of the track
          schema:
            type: string
      responses:
        '201':
          description: Track added successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. ID is not a valid UUID.
        '422':
          description: Track does not exist in the database.
      summary: Add a track to favorites
      tags: *ref_3
    delete:
      operationId: FavoritesController_removeTrack
      parameters:
        - name: id
          required: true
          in: path
          description: UUID of the track
          schema:
            type: string
      responses:
        '204':
          description: Track removed successfully.
        '400':
          description: Bad request. ID is not a valid UUID.
        '404':
          description: Track is not in favorites.
      summary: Remove a track from favorites
      tags: *ref_3
  /favs/album/{id}:
    post:
      operationId: FavoritesController_addAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: UUID of the album
          schema:
            type: string
      responses:
        '201':
          description: Album added successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. ID is not a valid UUID.
        '422':
          description: Album does not exist in the database.
      summary: Add an album to favorites
      tags: *ref_3
    delete:
      operationId: FavoritesController_removeAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: UUID of the album
          schema:
            type: string
      responses:
        '204':
          description: Album removed successfully.
        '404':
          description: Album is not in favorites.
      summary: Remove an album from favorites
      tags: *ref_3
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addArtist
      parameters:
        - name: id
          required: true
          in: path
          description: UUID of the artist
          schema:
            type: string
      responses:
        '201':
          description: Artist added successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. ID is not a valid UUID.
        '422':
          description: Artist does not exist in the database.
      summary: Add an artist to favorites
      tags: *ref_3
    delete:
      operationId: FavoritesController_removeArtist
      parameters:
        - name: id
          required: true
          in: path
          description: UUID of the artist
          schema:
            type: string
      responses:
        '204':
          description: Artist removed successfully.
        '404':
          description: Artist is not in favorites.
      summary: Remove an artist from favorites
      tags: *ref_3
  /user:
    post:
      operationId: UserController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SafeUserEntity'
        '400':
          description: >-
            Bad Request. body does not contain required fields or fields are
            invalid.
      summary: Create a new user
      tags: &ref_4
        - users
    get:
      operationId: UserController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SafeUserEntity'
      summary: Get all users
      tags: *ref_4
  /user/{id}:
    get:
      operationId: UserController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Returns single user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SafeUserEntity'
        '400':
          description: Bad Request. ID is invalid UUID.
        '404':
          description: User with ID not found.
      summary: Get user by id
      tags: *ref_4
    put:
      operationId: UserController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: User password updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SafeUserEntity'
        '400':
          description: >-
            Bad Request. Invalid ID, old password mismatch, or new password
            invalid.
        '404':
          description: User with ID not found.
      summary: Update user password
      tags: *ref_4
    delete:
      operationId: UserController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully (No Content returned)
        '400':
          description: Bad Request. ID is invalid UUID.
        '404':
          description: User with ID not found.
      summary: Delete user
      tags: *ref_4
info:
  title: Home Library Service API
  description: API documentation for the music library project
  version: '1.0'
  contact: {}
tags:
  - name: artists
    description: ''
  - name: albums
    description: ''
  - name: tracks
    description: ''
  - name: favorites
    description: ''
  - name: app
    description: ''
servers: []
components:
  schemas:
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the artist
          example: Freddie Mercury
        grammy:
          type: boolean
          description: Whether the artist has won a Grammy award
          example: true
      required:
        - name
        - grammy
    Artist:
      type: object
      properties:
        id:
          type: string
          description: The id  of the artist
          example: 4de888bd-af6a-497e-aace-f90641e9d33c
        name:
          type: string
          description: The name of the artist
          example: Freddie Mercury
        grammy:
          type: boolean
          description: Whether the artist has won a Grammy award
          example: true
      required:
        - id
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the artist
          example: Freddie Mercury
        grammy:
          type: boolean
          description: Whether the artist has won a Grammy award
          example: true
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the album
          example: Mutter
        year:
          type: number
          description: The release year of the album
          example: 2001
        artistId:
          type: string
          description: >-
            The UUID of the artist who created the album, or null if the artist
            does not exist.
          nullable: true
          example: &ref_5
            - 7f69ce5a-6eee-4fc4-a800-0372b05d9987
            - null
      required:
        - name
        - year
        - artistId
    Album:
      type: object
      properties:
        id:
          type: string
          description: The id  of the album
          example: 4de888bd-af6a-497e-aace-f90641e9d33c
        name:
          type: string
          description: The name of the album
          example: Mutter
        year:
          type: number
          description: The release year of the album
          example: 2001
        artistId:
          type: string
          description: >-
            The UUID of the artist who created the album, or null if the artist
            does not exist.
          nullable: true
          example:
            - 7f69ce5a-6eee-4fc4-a800-0372b05d9987
            - null
      required:
        - id
        - name
        - year
        - artistId
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the album
          example: Mutter
        year:
          type: number
          description: The release year of the album
          example: 2001
        artistId:
          type: string
          description: >-
            The UUID of the artist who created the album, or null if the artist
            does not exist.
          nullable: true
          example: *ref_5
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the track
          example: Nebel
        artistId:
          type: string
          description: >-
            The UUID of the artist who created the album, or null if the artist
            does not exist.
          nullable: true
          example: &ref_6
            - 7f69ce5a-6eee-4fc4-a800-0372b05d9987
            - null
        albumId:
          type: string
          description: >-
            The UUID of the album where the track belongs, or null if the album
            does not exist.
          nullable: true
          example: &ref_7
            - 7f69ce5a-6eee-4fc4-a800-0372b05d9987
            - null
        duration:
          type: number
          description: The duration of the track in sec
          example: 188
      required:
        - name
        - artistId
        - albumId
        - duration
    Track:
      type: object
      properties:
        id:
          type: string
          description: The id  of the track
          example: 4de888bd-af6a-497e-aace-f90641e9d33c
        name:
          type: string
          description: The name of the track
          example: Nebel
        artistId:
          type: string
          description: >-
            The UUID of the artist who created the album, or null if the artist
            does not exist.
          nullable: true
          example:
            - 7f69ce5a-6eee-4fc4-a800-0372b05d9987
            - null
        albumId:
          type: string
          description: >-
            The UUID of the album where the track belongs, or null if the album
            does not exist.
          nullable: true
          example:
            - 7f69ce5a-6eee-4fc4-a800-0372b05d9987
            - null
        duration:
          type: number
          description: The duration of the track in sec
          example: 188
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the track
          example: Nebel
        artistId:
          type: string
          description: >-
            The UUID of the artist who created the album, or null if the artist
            does not exist.
          nullable: true
          example: *ref_6
        albumId:
          type: string
          description: >-
            The UUID of the album where the track belongs, or null if the album
            does not exist.
          nullable: true
          example: *ref_7
        duration:
          type: number
          description: The duration of the track in sec
          example: 188
    FavoritesResponseDto:
      type: object
      properties:
        artists:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
      required:
        - artists
        - albums
        - tracks
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          description: The user login/username
          example: newUserLogin
        password:
          type: string
          description: 'The user password '
          example: P@ssw0rd123!
      required:
        - login
        - password
    SafeUserEntity:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the user (UUID)
          example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
          format: uuid
        login:
          type: string
          description: The user login/username
          example: johndoe123
        version:
          type: number
          description: Version of the user (incremented on each update)
          example: 1
        createdAt:
          type: number
          description: Timestamp of user creation (Unix epoch in milliseconds)
          example: 1678886400000
        updatedAt:
          type: number
          description: Timestamp of last user update (Unix epoch in milliseconds)
          example: 1678890000000
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          description: The user's current/old password
          example: CurrentP@ss123
        newPassword:
          type: string
          description: The user's new password
          example: NewAndBetterP@ss123
      required:
        - oldPassword
        - newPassword